refactor(scripts/i18n): extract addTranslationToBaseRules in order to be used in the website
